Refactored the files on routes folder and deleted unused files
